feat: 디테일 수정 및 로그인 로직 변경
