+++ ios/MOIM/main.m
@@ -5,6 +5,6 @@
 int main(int argc, char *argv[])
 {
   @autoreleasepool {
-    return UIApplicationMain(argc, argv, nil, NSStringFromClass([AppDelegate class]));
+      return UIApplicationMain(argc, argv, nil, NSStringFromClass([AppDelegate class]));
   }
 }
