+++ src/navigators/root/RootNavigator.tsx
@@ -4,9 +4,27 @@ import {useEffect} from 'react';
 import SplashScreen from 'react-native-splash-screen';
 import useAuth from 'hooks/queries/AuthScreen/useAuth.ts';
 import HomeStackNavigator from 'navigators/stack/HomeStackNavigator.tsx';
+import useFcmTokenStore from '../../stores/useFcmTokenStore.ts';
+import messaging from '@react-native-firebase/messaging';
 
 export default function RootNavigator() {
   const {isLogin, isLoginLoading} = useAuth();
+  const {setFcmToken} = useFcmTokenStore();
+
+  useEffect(() => {
+    async function getToken() {
+      try {
+        if (!messaging().isDeviceRegisteredForRemoteMessages) {
+          await messaging().registerDeviceForRemoteMessages();
+        }
+        const phoneToken = await messaging().getToken();
+        setFcmToken(phoneToken);
+      } catch (e) {
+        console.log(e);
+      }
+    }
+    getToken();
+  }, []);
 
   useEffect(() => {
     if (!isLoginLoading) {
